+++ benchflow-experiment-manager/application/src/main/java/cloud/benchflow/experimentmanager/BenchFlowExperimentManagerApplication.java
@@ -7,7 +7,7 @@ import cloud.benchflow.experimentmanager.services.external.DriversMakerService;
 import cloud.benchflow.experimentmanager.services.external.MinioService;
 import cloud.benchflow.experimentmanager.services.internal.dao.BenchFlowExperimentModelDAO;
 import cloud.benchflow.experimentmanager.services.internal.dao.TrialModelDAO;
-import cloud.benchflow.experimentmanager.tasks.ExperimentTaskScheduler;
+import cloud.benchflow.experimentmanager.scheduler.ExperimentTaskScheduler;
 import cloud.benchflow.faban.client.FabanClient;
 
 import com.mongodb.MongoClient;
